Fix AT-Robot arithmetic shifts
For both the SAL and SAR instructions, shifting zero positions
would result in all 0s for SAL and all 0s or 1s for SAR.  This
was because the sift amount was used to set the LOOP instruction
counter.  If zero, it would shift once, outright, then the LOOP
instruction would decrement CX to -1 and continue looping until
underflow eventually brought CX back to zero.  Needless to say
it always looped more times than bit positions leading to the
undesirable result.