Fixed aspect spike only resizes window if too small
On at least Linux machines, when the code reset the mode of the display,
the user's mouse would lose control focus on the window, preventing
them from continuing to resize.  This meant even when trying to make the
window bigger, they expand the dimenesions by a pixel or two and lose
mouse control of the resize.  This commit avoids this by only having the
code force the reset of mode when making the window smaller than the
original dimensions.